requirements.txt
Faker==20.1.0
psycopg2-binary==2.9.9
psycopg2_binary==2.9.9
python-decouple==3.8
setuptools==57.0.0
requests==2.31.0
geopandas==0.14.1
SQLAlchemy==2.0.23
multiprocess==0.70.15
GeoAlchemy2==0.14.2
xmltodict==0.13.0